requirements.txt
also==0.0.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.4
django-filter==0.12.0
django-guardian==1.3.2
django-toolbelt==0.0.1
django-uuidfield==0.5.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
google-api-python-client==1.4.2
gunicorn==19.4.5
httpie==0.9.3
httplib2==0.9.2
isodate==0.5.4
jsonfield==1.0.3
Markdown==2.6.5
oauth2client==1.5.2
pbr==1.8.1
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
Pygments==2.0.2
PyJWT==1.4.0
python-dateutil==2.4.2
requests==2.9.1
rsa==3.3
simplejson==3.8.1
six==1.10.0
static3==0.6.1
stevedore==1.10.0
underscore.py==0.1.6
uritemplate==0.6
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.24.0
